Change 2019.10 Part 1 from O(n^2) to O(n)
Previously we checked every occluding asteroid for every asteroid, which
is O(n^2). However, this isn't necessary - it's enough to calculate the
angles from the base station to every asteroid and group them by angles.
This will give you all of the asteroids in a line from the base station
together. If you need to know which is seen, it's then trivial to
calculate the one with the lowest manhattan distance to the base
station, but it's even easier to calculate the number of asteroids seen
- it's simply the number of different angles which asteroids can be seen
at, since if multiple asteroids are at a particular angle, only one of
them can be seen and will occlude all others.

As expected, this reduces the runtime significantly.