Simplify regex for 2015.05
As discovered when porting to Rust, the character class + backreference
combination here is not having the desired effect, but the requirement
is only to have two of the same characters with any character (including
the same) in the middle.

It looks like I originally interpreted this as only 'aba' but not 'aaa',
but because the character class did not do as intended I still got the
correct answer.