Add hacks::leak_string_to_str to allow comparing &'static str to String
Answers are known at compile time and are hard-coded in the binary, so
are &'static strs.
Our answers are generated by running the problem code, which produces a
String.

I haven't found a good way to compare this with the existing type
system. Ideally I'd like something like:

        type Part1Output = impl String

But this runs into issues around the types being different behind the
scenes.

Instead, this just leaks the String which allows the comparison. The
process is short lived so the memory is cleaned up almost immediately
after this comparison is made anyway.