Read the input file at runtime rather than embedding it in the binary
Although it's nice to have everything embedded in the binary, it comes
at the cost of an ugly boilerplate 'get_input' function at the top of
every solution.

This could likely be solved with a derive macro or something along those
lines, but the more straightforward solution is just to read the file at
runtime.