Make Captures::get* methods stateful
Each of these methods consumes a capture, removing the need to manually
specify the capture index at the call-site.

This removes the possibility of annoying copy-paste bugs like duplicate
indices, at the cost of having to call the methods in a specific order -
however it makes sense to call them in the order they appear in the
string anyway.